Add files via upload
Worked on Loan Approval Prediction. In this project, I have explored the data using the pandas function and cleared the null values and assigned meaningful values in their place. Used Machine Learning Algorithms to test the data and predict which model gives a better accuracy of the prediction value and made final analysis based on the same.